gestaltd: fold issue-95 remote apps into existing provider code
Merge gestalt remote provider into provider_client.go, error mapping into
provider_errors.go, and registration into provider_build.go. Drop the three
new files and redundant tests covered by public_grpc_test and
TestRemoteProviderRoundTrip.
