+++ gestaltd/internal/bootstrap/provider_build.go
@@ -91,7 +91,7 @@ func prepareProviderBuilds(
 		connAuth:       connAuth,
 		manualConnAuth: manualConnAuth,
 	}
-	if err := registerRemoteApps(builds.providers, cfg, deps.RemoteClients); err != nil {
+	if err := registerRemoteApps(builds.providers, cfg, deps); err != nil {
 		return nil, err
 	}
 	if len(cfg.Apps) == 0 {
@@ -124,6 +124,40 @@ func prepareProviderBuilds(
 		builds.pending = append(builds.pending, pendingProviderBuild{name: name, entry: entry, proxy: proxy, sha: sha})
 	}
 	return builds, nil
+}
+
+func registerRemoteApps(providers *registry.ProviderMap[core.Provider], cfg *config.Config, deps Deps) error {
+	clients := deps.RemoteClients
+	if providers == nil || cfg == nil || clients == nil || clients.App == nil {
+		return nil
+	}
+	for name, entry := range cfg.Apps {
+		name = strings.TrimSpace(name)
+		if name == "" || entry == nil || providerBuildsLocal(cfg, entry) {
+			continue
+		}
+		spec, _, err := buildStartupProviderSpec(name, entry)
+		if err != nil {
+			displayName := strings.TrimSpace(entry.DisplayName)
+			if displayName == "" {
+				displayName = name
+			}
+			spec = appservice.StaticProviderSpec{
+				Name:        name,
+				DisplayName: displayName,
+				Description: strings.TrimSpace(entry.Description),
+			}
+		}
+		provider := appservice.NewGestaltRemote(clients.App, spec)
+		if provider == nil {
+			return fmt.Errorf("remote app %q: provider client is required", name)
+		}
+		if err := providers.Register(name, provider); err != nil {
+			return fmt.Errorf("remote app %q: %w", name, err)
+		}
+		slog.Info("registered remote app provider", "provider", name)
+	}
+	return nil
 }
 
 func (b *preparedProviderBuilds) partition(categorize func(string, *config.ProviderEntry) AppStartupCategory) (noop, update *preparedProviderBuilds) {
